# chore(docs) - reduce build time