reduce build times by using string replace instead of JSDOM innerHTML.

Example from my system:

Before (14025ms):
```bash
> time npm run build

[11ty] Benchmark 14025ms 62%    85× (Configuration) "html-transform" Transform

[11ty] Copied 32 files / Wrote 85 files in 22.25 seconds (261.8ms each, v2.0.1)

npm run build  0.01s user 0.02s system 0% cpu 37.450 total
```

After (5021ms):
```bash
> time npm run build

[11ty] Benchmark   5021ms  39%    85× (Configuration) "html-transform" Transform

[11ty] Copied 32 files / Wrote 85 files in 12.75 seconds (150.0ms each, v2.0.1)

npm run build  0.01s user 0.02s system 0% cpu 28.081 total
```